create stacking effect on crypto cards
