+++ resources/js/Components/library/BaseCard.vue
@@ -19,7 +19,7 @@ const props = withDefaults(defineProps<Props>(), {
 <style scoped lang="scss">
 .card {
     border-radius: .5rem;
-    box-shadow: 0 1px 0.2rem 0 rgba(0, 0, 0, 0.1);
+    //box-shadow: 0 1px 0.2rem 0 rgba(0, 0, 0, 0.1);
     padding: 1rem;
 }
 </style>
